+++ src/main/java/net/thedevilthing/firstmod/datagen/ModDatapackProvider.java
@@ -1,0 +1,26 @@
+package net.thedevilthing.firstmod.datagen;
+
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.RegistrySetBuilder;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.PackOutput;
+import net.neoforged.neoforge.common.data.DatapackBuiltinEntriesProvider;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
+import net.thedevilthing.firstmod.Firstmod;
+import net.thedevilthing.firstmod.worldgen.ModBiomeModifiers;
+import net.thedevilthing.firstmod.worldgen.ModConfiguredFeatures;
+import net.thedevilthing.firstmod.worldgen.ModPlacedFeatures;
+
+import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+
+public class ModDatapackProvider extends DatapackBuiltinEntriesProvider {
+    public static final RegistrySetBuilder BUILDER = new RegistrySetBuilder()
+            .add(Registries.CONFIGURED_FEATURE, ModConfiguredFeatures::bootstrap)
+            .add(Registries.PLACED_FEATURE, ModPlacedFeatures::bootstrap)
+            .add(NeoForgeRegistries.Keys.BIOME_MODIFIERS, ModBiomeModifiers::bootstrap);
+
+    public ModDatapackProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> registries) {
+        super(output, registries, BUILDER, Set.of(Firstmod.MOD_ID));
+    }
+}