+++ src/main/java/net/thedevilthing/firstmod/worldgen/ModBiomeModifiers.java
@@ -1,0 +1,33 @@
+package net.thedevilthing.firstmod.worldgen;
+
+import net.minecraft.core.HolderSet;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.worldgen.BootstrapContext;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.BiomeTags;
+import net.minecraft.world.level.levelgen.GenerationStep;
+import net.neoforged.neoforge.common.world.BiomeModifier;
+import net.neoforged.neoforge.common.world.BiomeModifiers;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
+import net.thedevilthing.firstmod.Firstmod;
+
+import java.sql.Ref;
+
+public class ModBiomeModifiers {
+    public static final ResourceKey<BiomeModifier> ADD_NETHER_MERCURY_ORE = registerKey("add_nether_mercury_ore");
+
+    public static void bootstrap(BootstrapContext<BiomeModifier> context) {
+        var placedFeatures = context.lookup(Registries.PLACED_FEATURE);
+        var biomes = context.lookup(Registries.BIOME);
+
+        context.register(ADD_NETHER_MERCURY_ORE, new BiomeModifiers.AddFeaturesBiomeModifier(
+                biomes.getOrThrow(BiomeTags.IS_NETHER),
+                HolderSet.direct(placedFeatures.getOrThrow(ModPlacedFeatures.NETHER_MERCURY_ORE_PLACEED_KEY)),
+                GenerationStep.Decoration.UNDERGROUND_ORES));
+    }
+
+    private static ResourceKey<BiomeModifier> registerKey(String name) {
+        return ResourceKey.create(NeoForgeRegistries.Keys.BIOME_MODIFIERS, ResourceLocation.fromNamespaceAndPath(Firstmod.MOD_ID, name));
+    }
+}