+++ src/main/java/net/thedevilthing/firstmod/worldgen/ModConfiguredFeatures.java
@@ -1,0 +1,36 @@
+package net.thedevilthing.firstmod.worldgen;
+
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.worldgen.BootstrapContext;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.feature.Feature;
+import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.OreConfiguration;
+import net.minecraft.world.level.levelgen.structure.templatesystem.BlockMatchTest;
+import net.minecraft.world.level.levelgen.structure.templatesystem.RuleTest;
+import net.thedevilthing.firstmod.Firstmod;
+import net.thedevilthing.firstmod.block.ModBlocks;
+
+public class ModConfiguredFeatures {
+    public static final ResourceKey<ConfiguredFeature<?, ?>> NETHER_MERCURY_ORE_KEY = registerKey("nether_mercury_ore");
+
+    public static void bootstrap(BootstrapContext<ConfiguredFeature<?, ?>> context) {
+        RuleTest netherrackReplaceables = new BlockMatchTest(Blocks.NETHERRACK);
+
+        register(context, NETHER_MERCURY_ORE_KEY, Feature.ORE, new OreConfiguration(netherrackReplaceables,
+                ModBlocks.NETHER_MERCURY_ORE.get().defaultBlockState(), 5));
+    }
+
+    public static ResourceKey<ConfiguredFeature<?, ?>> registerKey(String name) {
+        return ResourceKey.create(Registries.CONFIGURED_FEATURE, ResourceLocation.fromNamespaceAndPath(Firstmod.MOD_ID, name));
+    }
+
+    private static <FC extends FeatureConfiguration, F extends Feature<FC>> void register(
+            BootstrapContext<ConfiguredFeature<?, ?>> context,
+            ResourceKey<ConfiguredFeature<?, ?>> key, F feature, FC configuration) {
+        context.register(key, new ConfiguredFeature<>(feature, configuration));
+    }
+}